Reindent all the files
See #162

This was done with the following Python snippet.

```
import json
import sys

for path in sys.argv[1:]:
    with open(path) as f:
        j = json.load(f)
    with open(path, "w") as f:
        json.dump(j, f, indent=2)
        f.write("\n")
```

Python was chosen because Python preserves insertion order in
dictionaries, so this lets us punt the question of what order to use for
the fields, until we have more involved tooling.

Two-space indent was chosen because almost all the existing files use a
two-space indent.

While almost all the existing files put a space before the ':' in
dictionaryies, this rewrites them to stick to the default. This matches
more standard JSON formatting and is more likely to be reproduced by
other tools.

A newline was added at the end because almost all of the existing files
have them, and this avoids tools complaining about a lack of a trailing
newline.